+++ source/extensions/isaacsim.gui.menu/isaacsim/gui/menu/tools_menu.py
@@ -80,7 +80,6 @@ class ToolsMenuExtension:
                                     MenuLayout.Item(name="Joint Position"),
                                     MenuLayout.Item(name="Joint Velocity"),
                                     MenuLayout.Item(name="Open Loop Gripper Controller"),
-                                    MenuLayout.Item(name="Surface Gripper"),
                                 ],
                             ),
                             MenuLayout.Item("ROS 2 OmniGraphs"),